ai/requirements.txt
fastapi
uvicorn
cohere
gitpython
radon

